+++ CRM/CRM.Server/Dto/Product/QueryForSaleData.cs
@@ -86,7 +86,7 @@ namespace CRM.Server.Dto.Product
         /// Asset qty
         /// 资产数量
         /// </summary>
-        public short? AssetQty { get; init; }
+        public int? AssetQty { get; init; }
 
         /// <summary>
         /// Currency
